refactor(server): remove legacy thread lifecycle RPCs
