+++ crates/harness-server/src/handlers/context.rs
@@ -8,7 +8,7 @@ use harness_context::{
     ContextItem,
 };
 use harness_core::run_id::RunIdentity;
-use harness_core::types::{Decision, Event, EventFilters, ProjectId, SessionId, ThreadId};
+use harness_core::types::{EventFilters, ThreadId};
 use harness_protocol::methods::{RpcResponse, INTERNAL_ERROR, NOT_FOUND};
 
 const CONTEXT_MANIFEST_HOOK: &str = "context_manifest";
@@ -64,72 +64,6 @@ pub async fn context_manifest_get(
     }
 }
 
-pub(crate) async fn record_context_composition(
-    state: &AppState,
-    thread_id: &ThreadId,
-    project: ProjectId,
-    task_profile: harness_context::TaskProfile,
-) -> Result<(), String> {
-    let mode = state.core.server.config.context.mode.into();
-    let request = ComposeRequest {
-        thread_id: thread_id.clone(),
-        run_id: current_run_id(),
-        project,
-        task_profile,
-        budget_hint: state.core.server.config.context.budget_tokens,
-    };
-    let composer = build_composer(state, mode).await;
-    let outcome = composer.compose(&request);
-    let (manifest, decision, reason, fatal_error) = match outcome {
-        Ok(composition) => {
-            let mut manifest = composition.manifest;
-            if mode == ComposeMode::Enforce {
-                let message = "context enforce mode is not available until composed injection replaces legacy injection paths";
-                manifest
-                    .warnings
-                    .push("context_enforce_not_wired".to_string());
-                tracing::error!(thread_id = %thread_id, "context enforce mode requested before injection wiring is available");
-                (
-                    manifest,
-                    Decision::Block,
-                    Some("context_enforce_not_wired".to_string()),
-                    Some(message.to_string()),
-                )
-            } else {
-                let decision =
-                    if manifest.provider_errors.is_empty() && manifest.warnings.is_empty() {
-                        Decision::Pass
-                    } else {
-                        Decision::Warn
-                    };
-                (manifest, decision, None, None)
-            }
-        }
-        Err(error) => {
-            tracing::error!(thread_id = %thread_id, error = %error, "context composition failed");
-            let fatal_error = if mode == ComposeMode::Enforce {
-                Some(error.to_string())
-            } else {
-                None
-            };
-            (
-                error.manifest().clone(),
-                Decision::Block,
-                Some("compose_error".to_string()),
-                fatal_error,
-            )
-        }
-    };
-
-    if let Err(error) = log_manifest_event(state, thread_id, manifest, decision, reason).await {
-        tracing::warn!(thread_id = %thread_id, error = %error, "context manifest logging failed");
-    }
-    if let Some(error) = fatal_error {
-        return Err(error);
-    }
-    Ok(())
-}
-
 async fn build_composer(state: &AppState, mode: ComposeMode) -> ContextComposer {
     let mut config = ComposeConfig::from(&state.core.server.config.context);
     config.mode = mode;
@@ -165,27 +99,6 @@ async fn build_composer(state: &AppState, mode: ComposeMode) -> ContextComposer 
             )))
         }
     }
-}
-
-async fn log_manifest_event(
-    state: &AppState,
-    thread_id: &ThreadId,
-    manifest: ComposeManifest,
-    decision: Decision,
-    reason: Option<String>,
-) -> anyhow::Result<()> {
-    let detail = serde_json::to_string(&manifest)?;
-    let mut event = Event::new(
-        SessionId::new(),
-        CONTEXT_MANIFEST_HOOK,
-        &thread_id.to_string(),
-        decision,
-    );
-    event.run_id = manifest.run_id.clone();
-    event.reason = reason;
-    event.detail = Some(detail);
-    state.observability.events.log(&event).await?;
-    Ok(())
 }
 
 fn composition_response(composition: Composition) -> serde_json::Value {
