+++ crates/harness-server/src/handlers/mod.rs
@@ -8,6 +8,7 @@ pub mod error;
 pub mod exec;
 pub mod gc;
 pub mod health;
+pub mod initialize;
 pub mod learn;
 pub mod observe;
 pub mod operator_monitor;
@@ -21,7 +22,6 @@ pub mod rules;
 pub mod runtime_hosts;
 pub mod runtime_project_cache;
 pub mod skills;
-pub mod thread;
 pub mod token_usage;
 pub mod usage_monitor;
 pub mod worktrees;
