+++ crates/harness-server/src/task_executor/helpers.rs
@@ -202,24 +202,8 @@ pub(crate) fn emit_runtime_notification(
 }
 
 #[cfg(test)]
-pub(crate) async fn persist_runtime_thread(
-    thread_db: &Option<crate::thread_db::ThreadDb>,
-    server: &crate::server::HarnessServer,
-    thread_id: &ThreadId,
-) {
-    if let Some(db) = thread_db {
-        if let Some(thread) = server.thread_manager.get_thread(thread_id) {
-            if let Err(err) = db.update(&thread).await {
-                tracing::warn!("thread_db persist failed during turn execution: {err}");
-            }
-        }
-    }
-}
-
-#[cfg(test)]
 pub(crate) async fn process_stream_item(
     server: &crate::server::HarnessServer,
-    thread_db: &Option<crate::thread_db::ThreadDb>,
     notify_tx: &Option<crate::notify::NotifySender>,
     notification_tx: &tokio::sync::broadcast::Sender<RpcNotification>,
     thread_id: &ThreadId,
@@ -233,8 +217,6 @@ pub(crate) async fn process_stream_item(
                 .add_item(thread_id, turn_id, item.clone())
             {
                 tracing::warn!("failed to append stream item_started to turn: {err}");
-            } else {
-                persist_runtime_thread(thread_db, server, thread_id).await;
             }
             emit_runtime_notification(
                 notify_tx,
@@ -251,8 +233,6 @@ pub(crate) async fn process_stream_item(
                 .add_item(thread_id, turn_id, item.clone())
             {
                 tracing::warn!("failed to append stream item_completed to turn: {err}");
-            } else {
-                persist_runtime_thread(thread_db, server, thread_id).await;
             }
             emit_runtime_notification(
                 notify_tx,
@@ -270,8 +250,6 @@ pub(crate) async fn process_stream_item(
                     .set_turn_token_usage(thread_id, turn_id, usage.clone())
             {
                 tracing::warn!("failed to update turn token usage: {err}");
-            } else {
-                persist_runtime_thread(thread_db, server, thread_id).await;
             }
             emit_runtime_notification(
                 notify_tx,
@@ -289,8 +267,6 @@ pub(crate) async fn process_stream_item(
                 harness_core::types::Item::Error { code: -1, message },
             ) {
                 tracing::warn!("failed to append stream error item to turn: {err}");
-            } else {
-                persist_runtime_thread(thread_db, server, thread_id).await;
             }
         }
         StreamItem::MessageDelta { text } => {
@@ -325,8 +301,6 @@ pub(crate) async fn process_stream_item(
                 },
             ) {
                 tracing::warn!("failed to append approval request item to turn: {err}");
-            } else {
-                persist_runtime_thread(thread_db, server, thread_id).await;
             }
             emit_runtime_notification(
                 notify_tx,
@@ -595,6 +569,10 @@ pub(crate) async fn persist_artifact(
 #[cfg(test)]
 #[path = "helpers_tests.rs"]
 mod tests;
+
+#[cfg(test)]
+#[path = "helpers_context_stream_tests.rs"]
+mod context_stream_tests;
 
 #[cfg(test)]
 #[path = "helpers_prompt_cap_tests.rs"]
